+++ packages/swapkit/api/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @swapkit/api
 
+## 1.0.0-rc.74
+
+### Patch Changes
+
+- [#819](https://github.com/thorswap/SwapKit/pull/819) [`f389990`](https://github.com/thorswap/SwapKit/commit/f389990a4a14566315fe353a464d916fc247ce89) Thanks [@towanTG](https://github.com/towanTG)! - Type fixes and base URL fix
+
+- Updated dependencies [[`f389990`](https://github.com/thorswap/SwapKit/commit/f389990a4a14566315fe353a464d916fc247ce89)]:
+  - @swapkit/helpers@1.0.0-rc.106
+
 ## 1.0.0-rc.73
 
 ### Patch Changes